+++ django_peeringdb/management/commands/pdb_sync.py
@@ -3,10 +3,10 @@ sync peeringdb tables
 """
 from __future__ import print_function
 
+import calendar
 import logging
 import re
 from optparse import make_option
-import time
 from twentyc.rpc import RestClient
 
 import django.core.exceptions
@@ -48,7 +48,15 @@ class Command(BaseCommand):
         make_option('--only',
             action='store',
             default=False,
-            help='only process this ixp (id)'),
+            help='only process this table'),
+        make_option('--id',
+            action='store',
+            default=0,
+            help='only process this id'),
+        make_option('--limit',
+            type=int,
+            default=0,
+            help="limit objects retrieved, retrieve all objects if 0 (default)"),
         )
 # progress
 # quiet
@@ -61,10 +69,17 @@ class Command(BaseCommand):
             kwargs['user'] = settings.SYNC_USERNAME
             kwargs['password'] = settings.SYNC_PASSWORD
 
+        self.log.debug("syncing from {}".format(settings.SYNC_URL))
         self.connect(settings.SYNC_URL, **kwargs)
 
         # get models if limited by config
-        tables = self.get_class_list(settings.SYNC_ONLY)
+        only = options.get('only', settings.SYNC_ONLY)
+        self.log.debug("only tables {}".format(only))
+
+        pk = options.get('id', 0)
+
+        tables = self.get_class_list(only)
+        limit = options.get("limit", 0)
 
         # disable auto now
         for model in tables:
@@ -74,14 +89,14 @@ class Command(BaseCommand):
                 if field.name == "updated":
                     field.auto_now = False
 
-        self.sync(tables)
+        self.sync(tables, pk, limit=limit)
 
     def connect(self, url, **kwargs):
         self.rpc = RestClient(url, **kwargs)
 
-    def sync(self, tables):
+    def sync(self, tables, pk=0, **kwargs):
         for cls in tables:
-            self.update_db(cls, self.get_objs(cls))
+            self.update_db(cls, self.get_objs(cls, pk=pk, **kwargs))
 
     def get_class_list(self, only=None):
         tables = []
@@ -95,19 +110,24 @@ class Command(BaseCommand):
     def get_since(self, cls):
         upd = cls.handleref.last_change()
         if upd:
-            return int(time.mktime(upd.timetuple()))
+            return int(calendar.timegm(upd.timetuple()))
         return 0
 
     def get_data(self, cls, since):
         return self.rpc.all(cls._handleref.tag, since=since)
 
     def get_objs(self, cls, **kwargs):
-        since = self.get_since(cls)
+        pk = int(kwargs.pop('pk', 0))
+        if pk:
+            self.log.debug("getting single id={}".format(pk))
+            data = self.rpc.all(cls._handleref.tag, id=pk, **kwargs)
+            print("%s==%d %d changed" % (cls._handleref.tag, pk, len(data)))
 
-        data = self.rpc.all(cls._handleref.tag, since=since, **kwargs)
-        #data = self.rpc.all(cls._handleref.tag, since=since, limit=20)
-        print("%s last update %s %d changed" % (cls._handleref.tag, str(since), len(data)))
-        #print(data)
+        else:
+            since = self.get_since(cls)
+            data = self.rpc.all(cls._handleref.tag, since=since, **kwargs)
+            print("%s last update %s %d changed" % (cls._handleref.tag, str(since), len(data)))
+
         return data
 
     def cls_from_tag(self, tag):
@@ -125,23 +145,29 @@ class Command(BaseCommand):
         """
         try:
             sync.sync_obj(cls, row)
-        except django.core.exceptions.ValidationError, inst:
-           # There were validation errors
-           for field, errlst in inst.error_dict.items():
-                # check if it was a relationship that doesnt exist locally
-                m = re.match(".+ with id (\d+) does not exist.+", str(errlst))
-                if m:
-                    print("%s.%s not found locally, trying to fetch object... " % (field, m.group(1)))
-                    # fetch missing object
-                    r = self.rpc.get(field, int(m.group(1)), depth=0)
 
-                    # sync missing object
-                    self._sync(self.cls_from_tag(field), r[0])
-                else:
-                    raise
+        except django.core.exceptions.ObjectDoesNotExist as e:
+            # thrown by subquery on single row
+            print(e)
+            raise
+
+        except django.core.exceptions.ValidationError as inst:
+            # There were validation errors
+            for field, errlst in inst.error_dict.items():
+                 # check if it was a relationship that doesnt exist locally
+                 m = re.match(".+ with id (\d+) does not exist.+", str(errlst))
+                 if m:
+                     print("%s.%s not found locally, trying to fetch object... " % (field, m.group(1)))
+                     # fetch missing object
+                     r = self.rpc.get(field, int(m.group(1)), depth=0)
+
+                     # sync missing object
+                     self._sync(self.cls_from_tag(field), r[0])
+                 else:
+                     raise
            
-           # try to sync initial object once more
-           sync.sync_obj(cls, row)
+            # try to sync initial object once more
+            sync.sync_obj(cls, row)
 
     def update_db(self, cls, data):
         print("data to be processed", len(data))