+++ tests/test_sync.py
@@ -10,13 +10,56 @@ from django.core.management import (
 )
 from django.test import TestCase
 import django_peeringdb.models
+from django_peeringdb import settings
+
+# set this to 0 to fetch all objects during sync test (takes forever)
+# setting this to anything > 0 will trigger missing object validation errors
+# causing the test to also test the fallback solution for that scenario
+FETCH_LIMIT = 3
 
 
+def sync_test(f):
+    """
+    check for sync settings, configure
+    return decorator for all tests that sync
+    """
+    sync = pytest.config.getoption("--sync")
+
+    # check explicitly for False, none means default (which is prod)
+    if sync == False:
+        return pytest.mark.skip(reason="need --sync option to run sync tests")(f)
+
+    f.sync_args = dict(
+        debug=pytest.config.getoption('--sync-debug'),
+        )
+
+    sync_only = pytest.config.getoption("--sync-only")
+    if sync_only:
+        f.sync_args['only'] = sync_only
+
+    sync_id = pytest.config.getoption("--sync-id")
+    if sync_id:
+        f.sync_args['id'] = sync_id
+
+    if not sync or sync == 'prod':
+        pass
+
+    elif sync == 'beta':
+        settings.SYNC_URL = 'https://beta.peeringdb.com/api'
+
+    else:
+        settings.SYNC_URL = sync
+
+
+    return f
+
+
+@sync_test
 class SyncTests(TestCase):
     """ test sync command """
 
     def setUp(self):
-        self.cmd = load_command_class('django_peeringdb', "pdb_sync")
+        self.cmd = load_command_class('django_peeringdb', 'pdb_sync')
 
     def test_get_since_empty(self):
         for cls in self.cmd.get_class_list():
@@ -25,7 +68,11 @@ class SyncTests(TestCase):
             assert 0 == self.cmd.get_since(cls)
 
     def test_sync_all(self):
-        self.cmd.handle()
+        kwargs = getattr(self, 'sync_args', {})
+        if "limit" not in kwargs:
+            kwargs["limit"] = FETCH_LIMIT
+        print("syncing kwargs {}".format(kwargs))
+        self.cmd.handle(**kwargs)
 #        self.cmd.sync(self.cmd.get_class_list())
 
 #        for cls in self.cmd.get_class_list():
@@ -47,4 +94,7 @@ class SyncTests(TestCase):
             assert 0 != self.cmd.get_since(cls)
             assert 0 != cls.objects.all().count()
             print(cls.objects.all().count())
+
+    def test_sync_call_command(self):
+        call_command('pdb_sync', interactive=False, limit=FETCH_LIMIT)